requirements.txt
python-telegram-bot==20.3
pymongo==4.6.1
dnspython==2.3.0